+++ src/main/java/bio/terra/cli/service/WorkspaceManagerServiceGcp.java
@@ -19,7 +19,6 @@ import bio.terra.cli.serialization.userfacing.input.UpdateReferencedBqTableParam
 import bio.terra.cli.serialization.userfacing.input.UpdateReferencedGcsBucketParams;
 import bio.terra.cli.serialization.userfacing.input.UpdateReferencedGcsObjectParams;
 import bio.terra.cli.service.utils.HttpUtils;
-import bio.terra.cli.service.utils.WsmUtils;
 import bio.terra.workspace.api.ControlledGcpResourceApi;
 import bio.terra.workspace.api.ReferencedGcpResourceApi;
 import bio.terra.workspace.model.CreateControlledGcpAiNotebookInstanceRequestBody;
@@ -164,8 +163,8 @@ public class WorkspaceManagerServiceGcp extends WorkspaceManagerService {
       GcpGcsBucketLifecycleRuleCondition condition =
           new GcpGcsBucketLifecycleRuleCondition()
               .age(rule.condition.age)
-              .createdBefore(WsmUtils.dateAtMidnightAndUTC(rule.condition.createdBefore))
-              .customTimeBefore(WsmUtils.dateAtMidnightAndUTC(rule.condition.customTimeBefore))
+              .createdBefore(dateAtMidnightAndUTC(rule.condition.createdBefore))
+              .customTimeBefore(dateAtMidnightAndUTC(rule.condition.customTimeBefore))
               .daysSinceCustomTime(rule.condition.daysSinceCustomTime)
               .daysSinceNoncurrentTime(rule.condition.daysSinceNoncurrentTime)
               .live(rule.condition.isLive)
@@ -173,8 +172,7 @@ public class WorkspaceManagerServiceGcp extends WorkspaceManagerService {
                   rule.condition.matchesStorageClass.stream()
                       .map(GcsStorageClass::toWSMEnum)
                       .collect(Collectors.toList()))
-              .noncurrentTimeBefore(
-                  WsmUtils.dateAtMidnightAndUTC(rule.condition.noncurrentTimeBefore))
+              .noncurrentTimeBefore(dateAtMidnightAndUTC(rule.condition.noncurrentTimeBefore))
               .numNewerVersions(rule.condition.numNewerVersions);
 
       GcpGcsBucketLifecycleRule lifecycleRuleRequestObject =
@@ -301,7 +299,7 @@ public class WorkspaceManagerServiceGcp extends WorkspaceManagerService {
     String jobId = UUID.randomUUID().toString();
     CreateControlledGcpAiNotebookInstanceRequestBody createRequest =
         new CreateControlledGcpAiNotebookInstanceRequestBody()
-            .common(WsmUtils.createCommonFields(createParams.resourceFields))
+            .common(createCommonFields(createParams.resourceFields))
             .aiNotebookInstance(fromCLIObject(createParams))
             .jobControl(new JobControl().id(jobId));
     logger.debug("Create controlled GCP notebook request {}", createRequest);
@@ -313,7 +311,7 @@ public class WorkspaceManagerServiceGcp extends WorkspaceManagerService {
           // Start the GCP notebook creation job.
           HttpUtils.callWithRetries(
               () -> controlledGcpResourceApi.createAiNotebookInstance(createRequest, workspaceId),
-              WsmUtils::isRetryable);
+              WorkspaceManagerService::isRetryable);
 
           // Poll the result endpoint until the job is no longer RUNNING.
           CreatedControlledGcpAiNotebookInstanceResult createResult =
@@ -321,14 +319,13 @@ public class WorkspaceManagerServiceGcp extends WorkspaceManagerService {
                   () ->
                       controlledGcpResourceApi.getCreateAiNotebookInstanceResult(
                           workspaceId, jobId),
-                  (result) -> WsmUtils.isDone(result.getJobReport()),
-                  WsmUtils::isRetryable,
+                  (result) -> isDone(result.getJobReport()),
+                  WorkspaceManagerService::isRetryable,
                   // Creating a GCP notebook instance should take less than ~10 minutes.
                   60,
                   Duration.ofSeconds(10));
           logger.debug("Create controlled GCP notebook result {}", createResult);
-          WsmUtils.throwIfJobNotCompleted(
-              createResult.getJobReport(), createResult.getErrorReport());
+          throwIfJobNotCompleted(createResult.getJobReport(), createResult.getErrorReport());
           return createResult.getAiNotebookInstance();
         },
         "Error creating controlled GCP Notebook instance in the workspace.");
@@ -351,7 +348,7 @@ public class WorkspaceManagerServiceGcp extends WorkspaceManagerService {
     // convert the CLI object to a WSM request object
     CreateControlledGcpGcsBucketRequestBody createRequest =
         new CreateControlledGcpGcsBucketRequestBody()
-            .common(WsmUtils.createCommonFields(createParams.resourceFields))
+            .common(createCommonFields(createParams.resourceFields))
             .gcsBucket(
                 new GcpGcsBucketCreationParameters()
                     .name(createParams.bucketName)
@@ -380,7 +377,7 @@ public class WorkspaceManagerServiceGcp extends WorkspaceManagerService {
     // convert the CLI object to a WSM request object
     CreateControlledGcpBigQueryDatasetRequestBody createRequest =
         new CreateControlledGcpBigQueryDatasetRequestBody()
-            .common(WsmUtils.createCommonFields(createParams.resourceFields))
+            .common(createCommonFields(createParams.resourceFields))
             .dataset(
                 new GcpBigQueryDatasetCreationParameters()
                     .datasetId(createParams.datasetId)
@@ -677,7 +674,7 @@ public class WorkspaceManagerServiceGcp extends WorkspaceManagerService {
               () ->
                   controlledGcpResourceApi.deleteAiNotebookInstance(
                       deleteRequest, workspaceId, resourceId),
-              WsmUtils::isRetryable);
+              WorkspaceManagerService::isRetryable);
 
           // poll the result endpoint until the job is no longer RUNNING
           DeleteControlledGcpAiNotebookInstanceResult deleteResult =
@@ -685,12 +682,11 @@ public class WorkspaceManagerServiceGcp extends WorkspaceManagerService {
                   () ->
                       controlledGcpResourceApi.getDeleteAiNotebookInstanceResult(
                           workspaceId, asyncJobId),
-                  (result) -> WsmUtils.isDone(result.getJobReport()),
-                  WsmUtils::isRetryable);
+                  (result) -> isDone(result.getJobReport()),
+                  WorkspaceManagerService::isRetryable);
           logger.debug("delete controlled GCP notebook instance result: {}", deleteResult);
 
-          WsmUtils.throwIfJobNotCompleted(
-              deleteResult.getJobReport(), deleteResult.getErrorReport());
+          throwIfJobNotCompleted(deleteResult.getJobReport(), deleteResult.getErrorReport());
           return true;
         },
         "Error deleting controlled GCP Notebook instance in the workspace.");
@@ -716,20 +712,19 @@ public class WorkspaceManagerServiceGcp extends WorkspaceManagerService {
           // make the initial delete request
           HttpUtils.callWithRetries(
               () -> controlledGcpResourceApi.deleteBucket(deleteRequest, workspaceId, resourceId),
-              WsmUtils::isRetryable);
+              WorkspaceManagerService::isRetryable);
 
           // poll the result endpoint until the job is no longer RUNNING
           DeleteControlledGcpGcsBucketResult deleteResult =
               HttpUtils.pollWithRetries(
                   () -> controlledGcpResourceApi.getDeleteBucketResult(workspaceId, asyncJobId),
-                  (result) -> WsmUtils.isDone(result.getJobReport()),
-                  WsmUtils::isRetryable,
+                  (result) -> isDone(result.getJobReport()),
+                  WorkspaceManagerService::isRetryable,
                   /*maxCalls=*/ 12,
                   /*sleepDuration=*/ Duration.ofSeconds(5));
           logger.debug("delete controlled gcs bucket result: {}", deleteResult);
 
-          WsmUtils.throwIfJobNotCompleted(
-              deleteResult.getJobReport(), deleteResult.getErrorReport());
+          throwIfJobNotCompleted(deleteResult.getJobReport(), deleteResult.getErrorReport());
           return true;
         },
         "Error deleting controlled GCS bucket in the workspace.");
